Fixes start logging if incorrect config
If the application can not find the `path.properties` file for whatever
reason, it won't even log the error because the `logger` fiesd is not
initialized yet to th moment when the `loadAppProperties` function
called.

Also, the log line is incorrect, it should contain `{}` to actually
allow logging of the missing path.

Also, it makes sense to log the absolute path instead of relative for
user to know where to look for it.